refactor simtest
