+++ ref-farming/tests/test_one_farm_two_farmer.rs
@@ -1,6 +1,8 @@
 use near_sdk_sim::{call, init_simulator, to_yocto, view};
 use near_sdk::json_types::{U128};
-use crate::common::*;
+use crate::common::actions::*;
+use crate::common::utils::*;
+use crate::common::views::*;
 
 mod common;
